+++ nuklear/handler.go
@@ -5,6 +5,7 @@ import (
 	"github.com/raedahgroup/godcr/app"
 	"github.com/raedahgroup/godcr/app/walletcore"
 	"github.com/raedahgroup/godcr/nuklear/pagehandlers"
+	"github.com/raedahgroup/godcr/nuklear/pagehandlers/transaction"
 	"github.com/raedahgroup/godcr/nuklear/styles"
 	"github.com/raedahgroup/godcr/nuklear/widgets"
 )
@@ -62,7 +63,7 @@ func getNavPages() []navPage {
 		{
 			name:    "history",
 			label:   "History",
-			handler: &pagehandlers.HistoryHandler{},
+			handler: &transaction.HistoryHandler{},
 		},
 		{
 			name:    "send",
